Version 2
-Switched to ET online news feed
-Economize on ChatGPT API calls by calling once per day
-Added original news link
