+++ README.md
@@ -1,4 +1,4 @@
 # GPT-News
-This is a Flask application that's a kind of mashup between a news reader and ChatGPT. I started with a Flask tutorial from Digital Ocean (https://www.digitalocean.com/community/tutorials/how-to-make-a-web-application-using-flask-in-python-3) and modified it so that instead of storing and displaying blog posts, we grab a news feed from CNN, ask ChatGPT to read it using the ChatGPT API, and then ask ChatGPT to use segue to a (hopefully funny) sales pitch for a product. The CSS and HTML files are mostly unchanged from the tutorial app. The .py file is different. 
+This is a Flask application that's a kind of mashup between a news reader and ChatGPT. I started with a Flask tutorial from Digital Ocean (https://www.digitalocean.com/community/tutorials/how-to-make-a-web-application-using-flask-in-python-3) and modified it so that instead of storing and displaying blog posts, we grab a news feed from CNN, ask ChatGPT to read it using the ChatGPT API, and then ask ChatGPT to use segue to a (hopefully funny) sales pitch for a product. The CSS and HTML files are very similar to those in the tutorial app. 
 
 You need to get an OpenAI API token and insert it into the code where indicated.
