+++ requirements.txt
@@ -1,0 +1,6 @@
+Flask>=2.0.2
+openai
+feedparser
+
+
+
