+++ static/css/style.css
@@ -4,3 +4,5 @@ h1 {
     text-align: center;
     padding: 10px;
 }
+
+
