+++ static/css/style_old.css
@@ -1,0 +1,20 @@
+h1 {
+    border: 2px #eee solid;
+    color: brown;
+    text-align: center;
+    padding: 10px;
+}
+
+h2 {
+    border: 2px #eee solid;
+    color: blue;
+    text-align: center;
+    padding: 10px;
+}
+
+h4 {
+    border: 2px #fee solid;
+    color: blue;
+    text-align: center;
+    padding: 10px;
+}
