+++ templates/about.html
@@ -1,0 +1,12 @@
+{% extends 'base.html' %}
+
+{% block content %}
+    <h1>About This Site</h1>
+    <p></p>
+    <p>This site is generated by looking at the day's top stories at Entertainment Weekly online, and calling Open AI's ChatGPT API using some of those stories as a prompt. I added the further prompt that we want ChatGPT to generate a sales pitch for a fun product such as the Ginsu Knife, or the Snuggie.</p>
+    <p>The idea was to create an example of a business use for Chat GPT, such as creating a topical sales pitch for a product.</p>
+    <p>I don't control the content that ChatGPT generates. If you read something that offends you or which is false (entirely possible), then go and try Elon Musk's "Truth GPT." I'm sure you'll like it better.</p>
+    <h1>About Mike Sadowski </h1>
+    <p></p>
+    <p>I’m head of Application Development at a major publicly-traded real-estate company. Previously I was CTO at a SaaS logistics software company, and CTO of a major logistics service provider. You can learn more about this project and my other side projects at <a href="http://www.mikesml.com/">my blog.</a></p>
+{% endblock %}
