+++ templates/base.html
@@ -21,7 +21,7 @@ aria-controls="navbarNav" aria-expanded="false" aria-label="Toggle navigation">
         <div class="collapse navbar-collapse" id="navbarNav">
             <ul class="navbar-nav">
             <li class="nav-item active">
-                <a class="nav-link" href="#">About</a>
+                <a class="nav-link" href="{{ url_for('about')}}">About</a>
             </li>
             </ul>
         </div>
