+++ templates/post.html
@@ -4,5 +4,8 @@
     <h2>{% block title %} {{ post[2] }} {% endblock %}</h2>
     <span class="badge badge-primary">{{ post[1] }}</span>
     <p>{{ post[3] }}</p>
+    <a target="_blank" rel="noopener noreferrer" href= {{post[5]}}>
+       <p>original story: {{ post[5] }}</p>
+    </a>
 {% endblock %}
